+++ Mods ChoGGi/Find Map Locations/metadata.lua
@@ -18,9 +18,9 @@ return PlaceObj("ModDef", {
 	"pops_any_uuid", "b97dbd4e-71cf-4e82-80a0-46dd104133bc",
 	"steam_id", "2453011286",
 	"lua_revision", 1007000, -- Picard
-	"version", 7,
+	"version", 8,
 	"version_major", 0,
-	"version_minor", 7,
+	"version_minor", 8,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {