+++ Mods ChoGGi/More Terrain Rocks/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "1743036391",
 	"pops_any_uuid", "22eb2462-5d89-415f-9e47-a7c0b45ef171",
 	"lua_revision", 1007000, -- Picard
-	"version", 5,
+	"version", 6,
 	"version_major", 0,
-	"version_minor", 5,
+	"version_minor", 6,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {